+++ src/main/kotlin/com/nanovms/ops/action/BuildAction.kt
@@ -1,25 +1,26 @@
-package com.nanovms.ops
+package com.nanovms.ops.action
 
-import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.service
 import com.intellij.openapi.fileChooser.FileChooser
 import com.intellij.openapi.fileChooser.FileChooserDescriptor
+import com.nanovms.ops.Log
+import com.nanovms.ops.OpsService
 
-class RunAction : BaseAction() {
+class BuildAction : BaseAction() {
     override fun actionPerformed(e: AnActionEvent) {
         val descriptor = FileChooserDescriptor(true, false, true, false, false, false)
         val selectedFiles = FileChooser.chooseFiles(descriptor, null, null)
         if (selectedFiles.isNotEmpty()) {
-            val filepath = selectedFiles[0].path
+            val filePath = selectedFiles[0].path
             val ops = service<OpsService>()
-            val result = ops.runExecutable(filepath)
-            if (result.hasError) {
-                Log.error(result.error)
-                Log.notifyError("Failed to execute $filepath")
+            val result = ops.build(filePath)
+            if(result.hasError) {
+                Log.notifyError("Failed to build image for '${filePath}'")
+                result.process?.let { Log.error(it.errorOutput) }
                 return
             }
-            Log.infoAndNotify("Executed $filepath with PID #${result.process?.pid}")
+            Log.notifyInfo("Built image for '${filePath}'")
         }
     }
 }
