+++ src/main/kotlin/com/nanovms/ops/action/RunAction.kt
@@ -1,24 +1,20 @@
-package com.nanovms.ops
+package com.nanovms.ops.action
 
-import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.service
 import com.intellij.openapi.fileChooser.FileChooser
 import com.intellij.openapi.fileChooser.FileChooserDescriptor
+import com.nanovms.ops.Log
+import com.nanovms.ops.OpsService
 
-class RunWithConfigAction : BaseAction() {
+class RunAction : BaseAction() {
     override fun actionPerformed(e: AnActionEvent) {
         val descriptor = FileChooserDescriptor(true, false, true, false, false, false)
-        descriptor.title = "Select ELF or Javascript File"
-        var selectedFiles = FileChooser.chooseFiles(descriptor, null, null)
+        val selectedFiles = FileChooser.chooseFiles(descriptor, null, null)
         if (selectedFiles.isNotEmpty()) {
             val filepath = selectedFiles[0].path
-
-            descriptor.title = "Select Configuration File"
-            selectedFiles = FileChooser.chooseFiles(descriptor, null, null)
-
             val ops = service<OpsService>()
-            val result = ops.runExecutable(filepath, selectedFiles[0].path)
+            val result = ops.runExecutable(filepath)
             if (result.hasError) {
                 Log.error(result.error)
                 Log.notifyError("Failed to execute $filepath")
