+++ src/main/kotlin/com/nanovms/ops/action/RunWithConfigAction.kt
@@ -1,0 +1,31 @@
+package com.nanovms.ops.action
+
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.components.service
+import com.intellij.openapi.fileChooser.FileChooser
+import com.intellij.openapi.fileChooser.FileChooserDescriptor
+import com.nanovms.ops.Log
+import com.nanovms.ops.OpsService
+
+class RunWithConfigAction : BaseAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        val descriptor = FileChooserDescriptor(true, false, true, false, false, false)
+        descriptor.title = "Select ELF or Javascript File"
+        var selectedFiles = FileChooser.chooseFiles(descriptor, null, null)
+        if (selectedFiles.isNotEmpty()) {
+            val filepath = selectedFiles[0].path
+
+            descriptor.title = "Select Configuration File"
+            selectedFiles = FileChooser.chooseFiles(descriptor, null, null)
+
+            val ops = service<OpsService>()
+            val result = ops.runExecutable(filepath, selectedFiles[0].path)
+            if (result.hasError) {
+                Log.error(result.error)
+                Log.notifyError("Failed to execute $filepath")
+                return
+            }
+            Log.infoAndNotify("Executed $filepath with PID #${result.process?.pid}")
+        }
+    }
+}
